requirements.txt
pandas~=1.3.5
praw~=7.5.0
scikit-learn~=1.0.2
python-decouple~=3.6
joblib~=1.1.0
numpy~=1.19.5